Update main.py
Fixed a bug that would inform the user that their email has already been registered when this is untrue. This is due to the missing scalar() after querying for the existing user under /register which returns a ChunkedIteratorResult object. Adding scalar() returns a NoneType object instead and allows code to flow without issue.